chore(build): refactor broccoli trees to generate them in order to reduce duplication
