+++ gulpfile.js
@@ -34,6 +34,7 @@ var util = require('./tools/build/util');
 var bundler = require('./tools/build/bundle');
 var replace = require('gulp-replace');
 var insert = require('gulp-insert');
+var makeBroccoliTree = require('./tools/broccoli/make-broccoli-tree');
 
 // Note: when DART_SDK is not found, all gulp tasks ending with `.dart` will be skipped.
 
@@ -756,11 +757,11 @@ gulp.task('build.dart', function(done) {
 });
 
 gulp.task('broccoli.js.dev', function() {
-  return broccoliBuild(require('./Brocfile-js_dev.js'), path.join('js', 'dev'));
+  return broccoliBuild(makeBroccoliTree('dev'), path.join('js', 'dev'));
 });
 
 gulp.task('broccoli.js.prod', function() {
-  return broccoliBuild(require('./Brocfile-js_prod.js'), path.join('js', 'prod'));
+  return broccoliBuild(makeBroccoliTree('prod'), path.join('js', 'prod'));
 });
 
 gulp.task('build.js.dev', function(done) {
@@ -774,7 +775,7 @@ gulp.task('build.js.dev', function(done) {
 gulp.task('build.js.prod', ['broccoli.js.prod']);
 
 gulp.task('broccoli.js.cjs', function() {
-  return broccoliBuild(require('./Brocfile-js_cjs.js'), path.join('js', 'cjs'));
+  return broccoliBuild(makeBroccoliTree('cjs'), path.join('js', 'cjs'));
 });
 gulp.task('build.js.cjs', function(done) {
   runSequence(
